requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
pluggy==1.3.0
psycopg==3.1.12
pytest==7.4.3
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing-extensions==4.8.0
urllib3==2.0.7